refactor(site): move AI budget display into the groups module
`AIBudgetAmount` and` AIBudgetUsage` are group-table spend UI, not shared primitives. Move them next to the existing groups helpers and keep `#/modules/groups` working via index.ts.
